Now it Automatic delte A users kvitteringer
