+++ ReciptHelper/app/modules/ProtectedRoute.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { Navigate } from "react-router-dom";
 
 const ProtectedRoute = ({ children }: { children: React.ReactNode }) => {
-  const isAuthenticated = Boolean(sessionStorage.getItem("email"));
+  const isAuthenticated = Boolean(localStorage.getItem("email"));
   
   if (!isAuthenticated) {
     return <Navigate to="/" replace />;
